+++ src/lintrans/__init__.py
@@ -8,6 +8,6 @@
 
 from . import crash_reporting, global_settings, gui, matrices, typing_, updating
 
-__version__ = '0.3.1-alpha'
+__version__ = '0.4.0-alpha'
 
 __all__ = ['crash_reporting', 'global_settings', 'gui', 'matrices', 'typing_', 'updating', '__version__']