+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - You can now use anonymous matrices (using numerical entries directly without needing to define a
   named matrix). See the tutorial for details
 - The input and output vectors now appear in the visual definition dialog, and are translucent
+- You are now always prompted to save the session upon exiting, even if you hadn't already saved
+- You can now edit the size of draggable points and the snapping distance for integer coordinates
 
 ### Fixed