+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.4.0] - 2022-11-17
 
 ### Added
 
@@ -121,7 +121,8 @@ instead, I'll just summarize the features.
 - Ability to zoom in and out of the viewport
 - Add dialog to change display settings
 
-[Unreleased]: https://github.com/DoctorDalek1963/lintrans/compare/v0.3.0...main
+[Unreleased]: https://github.com/DoctorDalek1963/lintrans/compare/v0.4.0...main
+[0.4.0]: https://github.com/DoctorDalek1963/lintrans/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/DoctorDalek1963/lintrans/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/DoctorDalek1963/lintrans/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/DoctorDalek1963/lintrans/compare/v0.2.0...v0.2.1